+++ lib/src/option.dart
@@ -4,24 +4,106 @@
 import 'package:equatable/equatable.dart';
 import './result.dart';
 
-/// The type of the option, either `some` or `none`, useful with `switch`.
-enum OptionType { some, none }
-
 /// Option is a type that represents either some value (`Some`) or none
 /// (`None`).
 ///
-/// `Option<Some>` is the type used for returning an optional value. It is an
-/// object with a `some` value, and `none`, representing no value.
-class Option<Some> extends Equatable {
-  final Some _some;
+/// `Option<T>` is the type used for returning an optional value. It is an
+/// object with a `Some` value, and `None`, representing no value.
+abstract class Option<T> extends Equatable {
+  Option();
 
   /// Create a `Some` option with the given value.
   ///
   /// Passing a `null` value will result in a `None`.
-  Option.some(Some v) : _some = v;
+  factory Option.some(T v) {
+    return v == null ? None() : Some(v);
+  }
 
   /// Create a `None` option with no value.
-  Option.none() : _some = null;
+  factory Option.none() => None();
+
+  /// Invokes either the `someop` or the `noneop` depending on the option.
+  ///
+  /// This is an attempt at providing something similar to the Rust `match`
+  /// expression, which makes it easy to handle both cases at once.
+  ///
+  /// See also [when] for another way to achieve the same behavior.
+  void match(void Function(T) someop, void Function() noneop);
+
+  /// Invokes either `some` or `none` depending on the option.
+  ///
+  /// Identical to [match] except that the arguments are named.
+  void when({void Function(T) some, void Function() none});
+
+  /// Unwraps an option, yielding the content of a `Some`.
+  ///
+  /// Throws an `Exception` if the value is a `None`, with the passed message.
+  T expect(String msg);
+
+  /// Unwraps an option, yielding the content of a `Some`.
+  ///
+  /// Throws an empty exception if this result is a `None`.
+  T unwrap();
+
+  /// Returns the contained value or a default.
+  T unwrapOr(T opt);
+
+  /// Returns the contained value or computes it from a closure.
+  T unwrapOrElse(T Function() op);
+
+  /// Maps an `Option<T>` to `Option<U>` by applying a function to a contained
+  /// `Some` value. Otherwise returns a `None`.
+  Option<U> map<U>(U Function(T) op);
+
+  /// Applies a function to the contained value (if any), or returns the
+  /// provided default (if not).
+  U mapOr<U>(U Function(T) op, U opt);
+
+  /// Maps an `Option<T>` to `U` by applying a function to a contained `T`
+  /// value, or computes a default (if not).
+  U mapOrElse<U>(U Function(T) op, U Function() def);
+
+  /// Transforms the `Option<T>` into a `Result<T, E>`, mapping `Some(v)` to
+  /// `Ok(v)` and `None` to `Err(err)`.
+  Result<T, E> okOr<E>(E err);
+
+  /// Transforms the `Option<T>` into a `Result<T, E>`, mapping `Some(v)` to
+  /// `Ok(v)` and `None` to `Err(err())`.
+  Result<T, E> okOrElse<E>(E Function() err);
+
+  /// Returns `None` if the option is `None`, otherwise calls `predicate` with
+  /// the wrapped value and returns:
+  ///
+  /// * `Some(t)` if predicate returns `true` (where `t` is the wrapped value)
+  /// * `None` if predicate returns `false`.
+  Option<T> filter(bool Function(T) predicate);
+
+  /// Returns `None` if the option is `None`, otherwise returns `optb`.
+  Option<T> and(Option<T> optb);
+
+  /// Returns `None` if the option is `None`, otherwise calls `op` with the
+  /// wrapped value and returns the result.
+  Option<T> andThen(Option<T> Function(T) op);
+
+  /// Returns the option if it contains a value, otherwise returns `optb`.
+  Option<T> or(Option<T> optb);
+
+  /// Returns the option if it contains a value, otherwise calls `op` and
+  /// returns the result.
+  Option<T> orElse(Option<T> Function() op);
+
+  /// Returns `Some` if exactly one of `this`, `optb` is `Some`, otherwise
+  /// returns `None`.
+  Option<T> xor(Option<T> optb);
+}
+
+class Some<T> extends Option<T> {
+  final T _some;
+
+  /// Create a `Some` option with the given value.
+  Some(v)
+      : assert(v != null),
+        _some = v;
 
   @override
   List<Object> get props => [_some];
@@ -30,186 +112,127 @@ class Option<Some> extends Equatable {
   bool get stringify => true;
 
   @override
-  bool operator ==(other) => other is Option && other._some == _some;
+  bool operator ==(other) => other is Some && other._some == _some;
 
-  /// Return the type of this option, either `some` or `none`.
-  OptionType type() {
-    return _some != null ? OptionType.some : OptionType.none;
+  @override
+  void match(void Function(T) someop, void Function() noneop) => someop(_some);
+
+  @override
+  void when({void Function(T) some, void Function() none}) => some(_some);
+
+  @override
+  T expect(String msg) => _some;
+
+  @override
+  T unwrap() => _some;
+
+  @override
+  T unwrapOr(T opt) => _some;
+
+  @override
+  T unwrapOrElse(T Function() op) => _some;
+
+  @override
+  Option<U> map<U>(U Function(T) op) => Option.some(op(_some));
+
+  @override
+  U mapOr<U>(U Function(T) op, U opt) => op(_some);
+
+  @override
+  U mapOrElse<U>(U Function(T) op, U Function() def) => op(_some);
+
+  @override
+  Result<T, E> okOr<E>(E err) => Result.ok(_some);
+
+  @override
+  Result<T, E> okOrElse<E>(E Function() err) => Result.ok(_some);
+
+  @override
+  Option<T> filter(bool Function(T) predicate) {
+    return predicate(_some) ? this : Option.none();
   }
 
-  /// Returns `true` if the option is `Some`.
-  bool get isSome => _some != null;
+  @override
+  Option<T> and(Option<T> optb) => optb;
 
-  /// Returns `true` if the option is `None`.
-  bool get isNone => _some == null;
+  @override
+  Option<T> andThen(Option<T> Function(T) op) => op(_some);
 
-  /// Invokes either the `someop` or the `noneop` depending on the option.
-  ///
-  /// This is an attempt at providing something similar to the Rust `match`
-  /// expression, which makes it easy to handle both cases at once.
-  ///
-  /// See also [when] for another way to achieve the same behavior.
-  void match(void Function(Some) someop, void Function() noneop) {
-    if (_some != null) {
-      someop(_some);
-    } else {
-      noneop();
-    }
+  @override
+  Option<T> or(Option<T> optb) => this;
+
+  @override
+  Option<T> orElse(Option<T> Function() op) => this;
+
+  @override
+  Option<T> xor(Option<T> optb) => optb is None ? this : Option.none();
+}
+
+class None<T> extends Option<T> {
+  /// Create a `None` option with no value.
+  None();
+
+  @override
+  List<Object> get props => [];
+
+  @override
+  bool get stringify => true;
+
+  @override
+  bool operator ==(other) => other is None;
+
+  @override
+  void match(void Function(T) someop, void Function() noneop) => noneop();
+
+  @override
+  void when({void Function(T) some, void Function() none}) => none();
+
+  @override
+  T expect(String msg) {
+    throw Exception(msg);
   }
 
-  /// Invokes either `some` or `none` depending on the option.
-  ///
-  /// Identical to [match] except that the arguments are named.
-  void when({void Function(Some) some, void Function() none}) {
-    if (_some != null) {
-      some(_some);
-    } else {
-      none();
-    }
+  @override
+  T unwrap() {
+    throw Exception();
   }
 
-  /// Unwraps an option, yielding the content of a `Some`.
-  ///
-  /// Throws an `Exception` if the value is a `None`, with the passed message.
-  Some expect(String msg) {
-    if (_some != null) {
-      return _some;
-    } else {
-      throw Exception(msg);
-    }
-  }
+  @override
+  T unwrapOr(T opt) => opt;
 
-  /// Unwraps an option, yielding the content of a `Some`.
-  ///
-  /// Throws an empty exception if this result is a `None`.
-  Some unwrap() {
-    if (_some != null) {
-      return _some;
-    } else {
-      throw Exception();
-    }
-  }
+  @override
+  T unwrapOrElse(T Function() op) => op();
 
-  /// Returns the contained value or a default.
-  Some unwrapOr(Some opt) {
-    return _some ?? opt;
-  }
+  @override
+  Option<U> map<U>(U Function(T) op) => Option.none();
 
-  /// Returns the contained value or computes it from a closure.
-  Some unwrapOrElse(Some Function() op) {
-    return _some ?? op();
-  }
+  @override
+  U mapOr<U>(U Function(T) op, U opt) => opt;
 
-  /// Maps an `Option<Some>` to `Option<NewSome>` by applying a function to
-  /// a contained `Some` value. Otherwise returns a `None`.
-  Option<NewSome> map<NewSome>(NewSome Function(Some) op) {
-    if (_some != null) {
-      return Option.some(op(_some));
-    } else {
-      return Option.none();
-    }
-  }
+  @override
+  U mapOrElse<U>(U Function(T) op, U Function() def) => def();
 
-  /// Applies a function to the contained value (if any), or returns the
-  /// provided default (if not).
-  NewSome mapOr<NewSome>(NewSome Function(Some) op, NewSome opt) {
-    if (_some != null) {
-      return op(_some);
-    } else {
-      return opt;
-    }
-  }
+  @override
+  Result<T, E> okOr<E>(E err) => Result.err(err);
 
-  /// Maps an `Option<Some>` to `NewSome` by applying a function to a contained
-  /// `Some` value, or computes a default (if not).
-  NewSome mapOrElse<NewSome>(
-      NewSome Function(Some) op, NewSome Function() def) {
-    if (_some != null) {
-      return op(_some);
-    } else {
-      return def();
-    }
-  }
+  @override
+  Result<T, E> okOrElse<E>(E Function() err) => Result.err(err());
 
-  /// Transforms the `Option<Some>` into a `Result<Some, Err>`, mapping
-  /// `Some(v)` to `Ok(v)` and `None` to `Err(err)`.
-  Result<Some, Err> okOr<Err>(Err err) {
-    if (_some != null) {
-      return Result.ok(_some);
-    } else {
-      return Result.err(err);
-    }
-  }
+  @override
+  Option<T> filter(bool Function(T) predicate) => Option.none();
 
-  /// Transforms the `Option<Some>` into a `Result<Some, Err>`, mapping
-  /// `Some(v)` to `Ok(v)` and `None` to `Err(err())`.
-  Result<Some, Err> okOrElse<Err>(Err Function() err) {
-    if (_some != null) {
-      return Result.ok(_some);
-    } else {
-      return Result.err(err());
-    }
-  }
+  @override
+  Option<T> and(Option<T> optb) => this;
 
-  /// Returns `None` if the option is `None`, otherwise calls `predicate` with
-  /// the wrapped value and returns:
-  ///
-  /// * `Some(t)` if predicate returns `true` (where `t` is the wrapped value)
-  /// * `None` if predicate returns `false`.
-  Option<Some> filter(bool Function(Some) predicate) {
-    if (_some != null && predicate(_some)) {
-      return this;
-    }
-    return Option.none();
-  }
+  @override
+  Option<T> andThen(Option<T> Function(T) op) => this;
 
-  /// Returns `None` if the option is `None`, otherwise returns `optb`.
-  Option<Some> and(Option<Some> optb) {
-    if (_some != null) {
-      return optb;
-    } else {
-      return this;
-    }
-  }
+  @override
+  Option<T> or(Option<T> optb) => optb;
 
-  /// Returns `None` if the option is `None`, otherwise calls `op` with the
-  /// wrapped value and returns the result.
-  Option<Some> andThen(Option<Some> Function(Some) op) {
-    if (_some != null) {
-      return op(_some);
-    } else {
-      return this;
-    }
-  }
+  @override
+  Option<T> orElse(Option<T> Function() op) => op();
 
-  /// Returns the option if it contains a value, otherwise returns `optb`.
-  Option<Some> or(Option<Some> optb) {
-    if (_some != null) {
-      return this;
-    } else {
-      return optb;
-    }
-  }
-
-  /// Returns the option if it contains a value, otherwise calls `op` and
-  /// returns the result.
-  Option<Some> orElse(Option<Some> Function() op) {
-    if (_some != null) {
-      return this;
-    } else {
-      return op();
-    }
-  }
-
-  /// Returns `Some` if exactly one of `this`, `optb` is `Some`, otherwise
-  /// returns `None`.
-  Option<Some> xor(Option<Some> optb) {
-    if (_some != null && optb.isNone) {
-      return this;
-    }
-    if (_some == null && optb.isSome) {
-      return optb;
-    }
-    return Option.none();
-  }
+  @override
+  Option<T> xor(Option<T> optb) => optb is Some ? optb : Option.none();
 }